Reorganized messages by partitioning them based on their pourpose
