+++ src/main/java/it/unitn/ds1/Client.java
@@ -12,6 +12,8 @@ import akka.actor.Cancellable;
 import akka.actor.Props;
 
 import it.unitn.ds1.models.*;
+import it.unitn.ds1.models.administratives.StartMsg;
+import it.unitn.ds1.models.administratives.StopMsg;
 import it.unitn.ds1.models.crash_detection.*;
 import it.unitn.ds1.utils.Delays;
 import scala.concurrent.duration.Duration;
