+++ src/main/java/it/unitn/ds1/Replica.java
@@ -17,7 +17,16 @@ import akka.actor.Cancellable;
 import akka.actor.Props;
 
 import it.unitn.ds1.models.*;
+import it.unitn.ds1.models.administratives.JoinGroupMsg;
+import it.unitn.ds1.models.administratives.StartMsg;
 import it.unitn.ds1.models.crash_detection.*;
+import it.unitn.ds1.models.election.CoordinatorMsg;
+import it.unitn.ds1.models.election.ElectionMsg;
+import it.unitn.ds1.models.election.LostUpdatesMsg;
+import it.unitn.ds1.models.election.SynchronizationMsg;
+import it.unitn.ds1.models.update.WriteAckMsg;
+import it.unitn.ds1.models.update.WriteMsg;
+import it.unitn.ds1.models.update.WriteOkMsg;
 import it.unitn.ds1.utils.Delays;
 import it.unitn.ds1.utils.UpdateRequestId;
 import it.unitn.ds1.utils.WriteId;
