+++ src/main/java/it/unitn/ds1/models/administratives/StopMsg.java
@@ -1,4 +1,4 @@
-package it.unitn.ds1.models;
+package it.unitn.ds1.models.administratives;
 
 import java.io.Serializable;
