+++ src/test/java/it/unitn/ds1/ElectionTest.java
@@ -3,6 +3,10 @@ package it.unitn.ds1;
 import akka.actor.ActorRef;
 import akka.actor.ActorSystem;
 import it.unitn.ds1.models.*;
+import it.unitn.ds1.models.administratives.JoinGroupMsg;
+import it.unitn.ds1.models.administratives.StartMsg;
+import it.unitn.ds1.models.election.ElectionMsg;
+import it.unitn.ds1.models.update.WriteMsg;
 import it.unitn.ds1.utils.WriteId;
 
 import java.io.IOException;
